Reformular o protocolo de ViewCode
Reformulando o protocolo e criando uma classe base para ser usada
