--- Hortali-App/Shared/Design System/Components/CollectionGroup/CollectionGroupStyle.swift
@@ -1,12 +1,0 @@
-/* Macro - Grupo 05 */
-
-
-/// Estilos do grupo de collection
-public enum CollectionGroupStyle {
-    
-    /// Mostra apenas a collection
-    case justCollection
-    
-    /// Mostra a collection com o título
-    case complete
-}
