+++ Hortali-App/Shared/Design System/ViewCode/ViewCode.swift
@@ -1,0 +1,60 @@
+/* Macro - Grupo 05 */
+
+/* Bibliotecas necessárias: */
+import UIKit
+
+
+/// Classe genérica para ser usada como herança nas Views
+class ViewCode: UIView, ViewCodeProtocol {
+    
+    /* MARK: - Atributos */
+    
+    var dynamicConstraints: [NSLayoutConstraint] = []
+    
+    
+    
+    /* MARK: - Construtor */
+    
+    init() {
+        super.init(frame: .zero)
+        self.createView()
+    }
+    
+    required init?(coder: NSCoder)  { fatalError("init(coder:) has not been implemented") }
+    
+    
+    
+    /* MARK: - Ciclo de Vida */
+    
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        self.reloadView()
+    }
+    
+    
+    
+    /* MARK: - Protocolo */
+    
+    /* ViewCodeProtocol */
+    
+    func setupHierarchy() { }
+    
+    func setupView() { }
+    
+    func setupStaticTexts() { }
+    
+    func setupUI() { }
+    
+    func setupFonts() { }
+    
+    func createDynamicConstraints() { }
+    
+    func createStaticConstraints() -> [NSLayoutConstraint] {
+        return []
+    }
+    
+    func clearDynamicConstraints() {
+        NSLayoutConstraint.deactivate(self.dynamicConstraints)
+        self.dynamicConstraints.removeAll()
+    }
+}
